REFACTOR: moved schema classes to seperate files
