+++ packages/server/src/schemas/index.ts
@@ -1,0 +1,7 @@
+export * from "./Tile";
+export * from "./Bomb";
+export * from "./Player";
+export * from "./Explosion";
+export * from "./PowerUp";
+export * from "./BaseTile";
+export * from "./GameState";
